# AI Optimized BPE trainer and tokenizer