--- a/cs336_basics/tokenizer.py
+++ b/cs336_basics/tokenizer.py
@@ -41,6 +41,8 @@ class Tokenizer(object):
         chunks = re.split(f"({special_pattern})", text) if self._special_tokens else [text]
 
         for chunk in chunks:
+            if self._debug:
+                print(f"Pretokenizing chunk: {chunk!r}")
             if not chunk:  # Skip empty chunks
                 continue
             elif chunk in self._special_tokens: