--- a/cs336_basics/tokenizer.py
+++ b/cs336_basics/tokenizer.py
@@ -38,7 +38,7 @@ class Tokenizer(object):
         # Sort special tokens by length (descending) to match longest first for overlapping cases
         sorted_special_tokens = sorted(self._special_tokens, key=len, reverse=True)
         special_pattern = "|".join([re.escape(token) for token in sorted_special_tokens])
-        chunks = re.split(f"({special_pattern})", text)
+        chunks = re.split(f"({special_pattern})", text) if self._special_tokens else [text]
 
         for chunk in chunks:
             if not chunk:  # Skip empty chunks