--- a/uv.toml
+++ b/uv.toml
@@ -1,0 +1,3 @@
+[[index]]
+url = "https://pypi.org/simple"
+default = true