``` markdown
BPE optimizationTinyStores-Train benchmark (train to 500 vocab size):

CLI: `uv run scripts/benchmark_bpe.py --vocab-size 1000 --output-dir ./test_benchmark --use-optimization`


- Naive algo: **13.45s**
- V0 heap & cache based: **5.78s**
- V1 optimization: **1.19s**
    - From <https://github.com/jasonzou0/cs336-ass1/commit/66ecdacb024f972c7497dd28beef3aef8d81d66b>
    - V0 profiling result on OWT dataset: <https://github.com/jasonzou0/cs336-ass1/blob/bpe/owt_train/optimized_merge.v0.svg>
    - Claude reimplemented optimizations (made to v1): <https://github.com/jasonzou0/cs336-ass1/commit/66ecdacb024f972c7497dd28beef3aef8d81d66b> (asked claude to reimplement those suggestions):
        - 1. Inlined `_extract_pairs` logic - Eliminated 11M+ function calls by inlining pair extraction directly in `_remove_token_from_tracking_by_index` and `_add_new_token`: **5.96s**
        - 2. Batch processing for set operations - Instead of individual set.add() and set.discard() calls, the new `update_tokens_incremental` method: **0.79s!!!**
            - Collects all pair changes first
            - Batches updates to minimize set operations
            - Reduces the number of heap operations
        - 3. Optimized tuple creation in `_perform_merge_in_token`:
            - Pre-computes tuple length
            - Avoids unnecessary tuple comparisons
            - Uses direct element comparison instead of tuple creation: **0.71s**
- V2 optimization (inline `_perform_merge_in_token` and `_extract_pairs`): **1.13s**
    - From <https://github.com/jasonzou0/cs336-ass1/commit/bedf55f0c57aa1de2313c7d31354e612439c2bbe>
- V3 optimization (batch set operations): **0.68s**
    - From <https://github.com/jasonzou0/cs336-ass1/commit/3289229ad616b351bec112acb19509032be5c14c>
- V4 optimization (update merged tokens tuple in place): **0.62s**
    - From <https://github.com/jasonzou0/cs336-ass1/commit/6187f8a44099aef2bf7509a5716c0bd380b2e500>

OWT benchmark (500 vocab size):

CLI: `uv run scripts/benchmark_bpe.py --vocab-size 1000 --output-dir ./owt_train --use-optimization --input-dataset data0/owt_train.txt `


- Naive algo:
- V0 heap & cache based:
- [V1 optimization](https://github.com/jasonzou0/cs336-ass1/commit/66ecdacb024f972c7497dd28beef3aef8d81d66b):
- [V2 optimization](https://github.com/jasonzou0/cs336-ass1/commit/bedf55f0c57aa1de2313c7d31354e612439c2bbe) (inline `_perform_merge_in_token` and `_extract_pairs`): **243s**
- [V3 optimization](https://github.com/jasonzou0/cs336-ass1/commit/3289229ad616b351bec112acb19509032be5c14c) (batch set operations): **328s**
- V4 optimization (update merged tokens tuple in place): **218s**
- V5 (Directly apply V4 opt on top of V2): **197.25s**

Learnings:

1.  **Super useful to have a perf benchmarking tool and run cpu profiles**
2.  **Biggest improvement is batch based set operations. This is Claud’s explanation on why it works:**


\# Our elements are tuples of bytes: (byte1, byte2)

pair = (b'a', b'b')

self.pair_positions[pair].add(token_idx)    \# ~60-80ns

self.pair_positions[pair].discard(token_idx) \# ~50-70ns


Batch Operation Advantage


Individual Approach (1000 operations):


1000 operations × 60ns = 60,000ns = 0.06ms

+ Function call overhead × 1000 = ~20,000ns

+ Python interpreter overhead = ~10,000ns

Total: ~0.09ms


Batch Approach (single operation):


C-level bulk processing = ~5,000ns = 0.005ms

+ Single function call = ~20ns

Total: ~0.005ms


Speedup: ~18x faster for this example!


Summary


- Individual cost: ~50-100ns per add()/discard()

- Batch operations: ~5-10ns per element (amortized)

- Speedup factor: 10-20x improvement with batching

- Critical factor: Function call overhead and Python interpreter overhead dominate for small operations


This is why batching provides such dramatic performance improvements in tight loops like BPE training!

```